+++ lumex/examples/utility/example_utility_workflow.cpp
@@ -4,6 +4,7 @@
 
 using namespace lumex::core::utility::demangle;
 using namespace lumex::core::utility::numeric;
+using namespace lumex::core::utility::process;
 
 int
 main ()
@@ -14,8 +15,9 @@ main ()
   SafeComparator<unsigned int> received (50U);
   int const max_payload = 40;
   if (received.safe_compare (max_payload))
-    std::cout << "accept length\n";
+    std::cout << "accept length; pid=" << get_current_pid () << '\n';
   else
-    std::cout << "reject length; type=" << lumDemangle (unsigned int) << '\n';
+    std::cout << "reject length; type=" << lumDemangle (unsigned int)
+              << " pid=" << get_current_pid () << '\n';
   return 0;
 }